Refactor include paths to remove redundant 'container/' prefix and add new header files for data structures
