+++ include/zelix/external_string.h
