+++ include/zelix/owned_string.h
