+++ include/zelix/shared_ptr.h
@@ -28,7 +28,7 @@
 #include <tuple>
 #include <type_traits>
 
-#include "zelix/container/forward.h"
+#include "zelix//forward.h"
 #include "zelix/memory/array_resource.h"
 #include "zelix/memory/monotonic.h"
 #include "zelix/memory/resource.h"
