+++ include/zelix/stream.h
@@ -28,8 +28,8 @@
 //
 
 #pragma once
-#include "zelix/container/vector.h"
-#include "zelix/container/optional.h"
+#include "zelix//vector.h"
+#include "zelix//optional.h"
 
 namespace zelix::stl
 {
