+++ include/zelix/string_equality.h
