+++ include/zelix/unique_ptr.h
@@ -27,7 +27,7 @@
 
 #include <tuple>
 #include <variant>
-#include "zelix/container/forward.h"
+#include "zelix//forward.h"
 #include "zelix/memory/array_resource.h"
 #include "zelix/memory/monotonic.h"
 #include "zelix/memory/resource.h"
